+++ qrl/core/apiprotocol.py
@@ -5,6 +5,7 @@ import copy
 import decimal
 import json
 import time
+from operator import itemgetter
 
 import statistics
 from twisted.internet.protocol import Protocol, connectionDone
@@ -12,16 +13,17 @@ from twisted.internet.protocol import Protocol, connectionDone
 from pyqrllib.pyqrllib import bin2hstr
 from qrl.core import logger
 from qrl.core.helper import json_print_telnet
-from qrl.core.Transaction_subtypes import TX_SUBTYPE_TX, TX_SUBTYPE_COINBASE
+from qrl.core.Transaction_subtypes import TX_SUBTYPE_TX, TX_SUBTYPE_COINBASE, TX_SUBTYPE_STAKE
+from qrl.core.Transaction import SimpleTransaction, CoinBase, Transaction
 
 
 class ApiProtocol(Protocol):
     def __init__(self):
         self.api_list = [
-            'block_data', 'stats', 'ip_geotag', 'exp_win', 'txhash', 'address',
-            'empty', 'last_tx', 'last_unconfirmed_tx', 'stake_reveal_ones', 'last_block', 
-            'richlist', 'ping', 'stake_commits', 'stake_reveals', 'stakers', 'next_stakers',
-            'latency', 'balance'
+            'block_data', 'stats', 'ip_geotag', 'txhash', 'address',
+            'empty', 'last_tx', 'last_unconfirmed_tx', 'last_block',
+            'richlist', 'ping', 'stakers', 'next_stakers', 'latency',
+            'balance'
         ]
 
     def parse_cmd(self, data):
@@ -61,15 +63,12 @@ class ApiProtocol(Protocol):
             self.transport.write(json_print_telnet(error))
             return False
 
-        #my_cls = ApiProtocol()  # call the command from api_list directly
         api_call = getattr(self, data[1].lower())
 
         if len(data) < 3:
             json_txt = api_call()
-        # self.transport.write(api_call())
         else:
             json_txt = api_call(data[2])
-        # self.transport.write(api_call(data[2]))
 
         http_header_GET = ("HTTP/1.1 200 OK\r\n"
                            "Content-Type: application/json\r\n"
@@ -83,10 +82,6 @@ class ApiProtocol(Protocol):
         self.transport.write(bytes(http_header_GET + json_txt, 'utf-8'))
         return
 
-    def exp_win(self, data=None):
-        logger.info('<<< API expected winner call')
-        return self.factory.chain.exp_win(data)
-
     def ping(self, data=None):
         logger.info('<<< API network latency ping call')
         self.factory.pos.p2pFactory.ping_peers()  # triggers ping for all connected peers at timestamp now. after pong response list is collated. previous list is delivered.
@@ -98,44 +93,211 @@ class ApiProtocol(Protocol):
 
     def stakers(self, data=None):
         logger.info('<<< API stakers call')
-        return self.factory.chain.stakers(data)
+
+        stakers = {'status': 'ok',
+                   'stake_list': []}
+
+        for staker in self.factory.state.stake_validators_list.sv_list:
+            sv = self.factory.state.stake_validators_list.sv_list[staker]
+            tmp_stakers = {'address': sv.stake_validator,
+                           'balance': sv.balance / 100000000.00000000,
+                           'hash_terminator': [],
+                           'nonce': sv.nonce}
+
+            for i in range(len(sv.hashchain_terminators)):
+                tmp_stakers['hash_terminator'].append(bin2hstr(sv.hashchain_terminators[i]))
+
+            stakers['stake_list'].append(tmp_stakers)
+
+        return json_print_telnet(stakers)
 
     def next_stakers(self, data=None):
         logger.info('<<< API next_stakers call')
-        return self.factory.chain.next_stakers(data)
 
-    def stake_commits(self, data=None):
-        logger.info('<<< API stake_commits call')
-        return self.factory.chain.stake_commits(data)
+        next_stakers = {'status': 'ok',
+                        'stake_list': []}
 
-    def stake_reveals(self, data=None):
-        logger.info('<<< API stake_reveals call')
-        return self.factory.chain.stake_reveals(data)
+        for staker in self.factory.state.stake_validators_list.next_sv_list:
+            sv = self.factory.state.stake_validators_list.next_sv_list[staker]
+            tmp_stakers = {'address': sv.stake_validator,
+                           'balance': sv.balance / 100000000.00000000,
+                           'hash_terminator': [],
+                           'nonce': sv.nonce}
 
-    def stake_reveal_ones(self, data=None):
-        logger.info('<<< API stake_reveal_ones')
-        return self.factory.chain.stake_reveal_ones(data)
+
+            for i in range(len(sv.hashchain_terminators)):
+                tmp_stakers['hash_terminator'].append(bin2hstr(sv.hashchain_terminators[i]))
+
+            next_stakers['stake_list'].append(tmp_stakers)
+
+        return json_print_telnet(next_stakers)
 
     def richlist(self, data=None):
+        """
+        only feasible while chain is small..
+        :param data:
+        :return:
+        """
         logger.info('<<< API richlist call')
-        return self.factory.chain.richlist(data)
+
+        if not data:
+            data = 5
+
+        error = {'status': 'error', 'error': 'invalid argument', 'method': 'richlist', 'parameter': data}
+
+        try:
+            n = int(data)
+        except:
+            return json_print_telnet(error)
+
+        if n <= 0 or n > 20:
+            return json_print_telnet(error)
+
+        if not self.factory.state.state_uptodate(self.factory.chain.m_blockheight()):
+            return json_print_telnet({'status': 'error', 'error': 'leveldb failed', 'method': 'richlist'})
+
+        addr = self.factory.state.db.return_all_addresses()
+        richlist = sorted(addr, key=itemgetter(1), reverse=True)
+
+        rl = {'richlist': {}}
+
+        if len(richlist) < n:
+            n = len(richlist)
+
+        for rich in richlist[:n]:
+            rl['richlist'][richlist.index(rich) + 1] = {}
+            rl['richlist'][richlist.index(rich) + 1]['address'] = rich[0].decode()
+            rl['richlist'][richlist.index(rich) + 1]['balance'] = rich[1] / 100000000.000000000
+
+        rl['status'] = 'ok'
+
+        return json_print_telnet(rl)
 
     def last_block(self, data=None):
         logger.info('<<< API last_block call')
-        return self.factory.chain.last_block(data)
+
+        error = {'status': 'error', 'error': 'invalid argument', 'method': 'last_block', 'parameter': data}
+
+        if not data:
+            data = 1
+
+        try:
+            n = int(data)
+        except:
+            return json_print_telnet(error)
+
+        if n <= 0 or n > 20:
+            return json_print_telnet(error)
+
+        lb = []
+        beginning = self.factory.chain.height() - n
+        for blocknum in range(self.factory.chain.height(), beginning - 1, -1):
+            block = self.factory.chain.m_get_block(blocknum)
+            lb.append(block)
+
+        last_blocks = {'blocks': []}
+        i = 0
+        for block in lb[1:]:
+            i += 1
+            tmp_block = {'blocknumber': block.blockheader.blocknumber,
+                         'block_reward': block.blockheader.block_reward / 100000000.00000000,
+                         'blockhash': bin2hstr(block.blockheader.prev_blockheaderhash),
+                         'timestamp': block.blockheader.timestamp,
+                         'block_interval': lb[i - 1].blockheader.timestamp - block.blockheader.timestamp,
+                         'number_transactions': len(block.transactions)}
+
+            last_blocks['blocks'].append(tmp_block)
+
+        last_blocks['status'] = 'ok'
+
+        return json_print_telnet(last_blocks)
 
     def last_unconfirmed_tx(self, data=None):
         logger.info('<<< API last_unconfirmed_tx call')
-        return self.factory.chain.last_unconfirmed_tx(data)
+
+        addr = {'transactions': []}
+        error = {'status': 'error', 'error': 'invalid argument', 'method': 'last_tx', 'parameter': data}
+
+        if not data:
+            data = 1
+        try:
+            n = int(data)
+        except:
+            return json_print_telnet(error)
+
+        if n <= 0 or n > 20:
+            return json_print_telnet(error)
+
+        tx_num = len(self.factory.chain.transaction_pool)
+        while tx_num > 0:
+            tx_num -= 1
+            tx = self.factory.chain.transaction_pool[tx_num]
+            if tx.subtype != TX_SUBTYPE_TX:
+                continue
+            tmp_txn = {'txhash': bin2hstr(tx.txhash),
+                       'block': 'unconfirmed',
+                       'timestamp': 'unconfirmed',
+                       'amount': tx.amount / 100000000.000000000,
+                       'type': tx.subtype}
+
+            tmp_txn['type'] = Transaction.tx_id_to_name(tmp_txn['type'])
+            addr['transactions'].append(tmp_txn)
+
+        addr['status'] = 'ok'
+        return json_print_telnet(addr)
 
     def last_tx(self, data=None):
         logger.info('<<< API last_tx call')
-        return self.factory.chain.last_tx(data)
+
+        if not data:
+            data = 1
+
+        addr = {'transactions': []}
+
+        error = {'status': 'error', 'error': 'invalid argument', 'method': 'last_tx', 'parameter': data}
+
+        try:
+            n = int(data)
+        except:
+            return json_print_telnet(error)
+
+        if n <= 0 or n > 20:
+            return json_print_telnet(error)
+
+        try:
+            last_txn = self.factory.state.db.get('last_txn')
+        except Exception:
+            error['error'] = 'txnhash not found'
+            return json_print_telnet(error)
+
+        n = min(len(last_txn), n)
+        while n > 0:
+            n -= 1
+            tx_meta = last_txn[n]
+            tx = SimpleTransaction().json_to_transaction(tx_meta[0])
+            tmp_txn = {'txhash': bin2hstr(tx.txhash),
+                       'block': tx_meta[1],
+                       'timestamp': tx_meta[2],
+                       'amount': tx.amount / 100000000.000000000,
+                       'type': tx.subtype}
+
+            addr['transactions'].append(tmp_txn)
+            addr['status'] = 'ok'
+
+        return json_print_telnet(addr)
 
     def ip_geotag(self, data=None):
         logger.info('<<< API ip_geotag call')
         self.factory.pos.p2pFactory.ip_geotag_peers()
-        return self.factory.chain.ip_geotag(data)
+        ip = {'status': 'ok',
+              'ip_geotag': self.factory.chain.ip_list}
+
+        x = 0
+        for i in self.factory.chain.ip_list:
+            ip['ip_geotag'][x] = i
+            x += 1
+
+        return json_print_telnet(ip)
 
     def empty(self, data=None):
         error = {
@@ -144,6 +306,38 @@ class ApiProtocol(Protocol):
             'methods available': self.api_list
         }
         return json_print_telnet(error)
+
+    def reformat_block(self, block):
+        block.blockheader.block_reward /= 100000000.000000000
+        block.blockheader.fee_reward /= 100000000.000000000
+        block.blockheader.reveal_hash = bin2hstr(block.blockheader.reveal_hash)
+        block.blockheader.vote_hash = bin2hstr(block.blockheader.vote_hash)
+        block.blockheader.headerhash = bin2hstr(block.blockheader.headerhash)
+        block.blockheader.tx_merkle_root = bin2hstr(block.blockheader.tx_merkle_root)
+        block.blockheader.prev_blockheaderhash = bin2hstr(block.blockheader.prev_blockheaderhash)
+
+        for txn in block.transactions:
+            self.reformat_txn(txn)
+
+        return block
+
+    def reformat_txn(self, txn):
+        if txn.subtype in (TX_SUBTYPE_TX, TX_SUBTYPE_COINBASE):
+            txn.amount = txn.amount / 100000000.000000000
+        txn.txhash = bin2hstr(txn.txhash)
+        txn.pubhash = bin2hstr(txn.pubhash)
+        txn.signature = bin2hstr(txn.signature)
+        txn.PK = bin2hstr(txn.PK)
+        if txn.subtype == TX_SUBTYPE_STAKE:
+            txn.slave_public_key = bin2hstr(txn.slave_public_key)
+            if txn.first_hash:
+                txn.first_hash = bin2hstr(txn.first_hash)
+            for j in range(len(txn.hash)):
+                txn.hash[j] = bin2hstr(txn.hash[j])
+
+        txn.subtype = Transaction.tx_id_to_name(txn.subtype)
+
+        return txn
 
     def block_data(self, data=None):  # if no data = last block ([-1])			#change this to add error..
         error = {
@@ -154,32 +348,22 @@ class ApiProtocol(Protocol):
         }
         logger.info('<<< API block data call %s', data)
         if not data:
-            data = self.factory.chain.m_get_last_block()
-            data1 = copy.deepcopy(data)
-            data1.status = 'ok'
-            return json_print_telnet(data1)
+            data = self.factory.chain.height()
+
         try:
             int(data)  # is the data actually a number?
         except:
             return json_print_telnet(error)
+
         js_bk = self.factory.chain.m_get_block(int(data))
+
         if not js_bk:
             return json_print_telnet(error)
         else:
             js_bk1 = copy.deepcopy(js_bk)
             js_bk1.number_transactions = len(js_bk1.transactions)
             js_bk1.status = 'ok'
-            js_bk1.blockheader.block_reward /= 100000000.000000000
-            js_bk1.blockheader.fee_reward /= 100000000.000000000
-            js_bk1.blockheader.headerhash = bin2hstr(js_bk1.blockheader.headerhash)
-            js_bk1.blockheader.prev_blockheaderhash = bin2hstr(js_bk1.blockheader.prev_blockheaderhash)
-            i = 0
-            for txn in js_bk1.transactions[0:]:
-                if txn.subtype in (TX_SUBTYPE_TX, TX_SUBTYPE_COINBASE):
-                    js_bk1.transactions[i].amount = txn.amount / 100000000.000000000
-                    js_bk1.transactions[i].txhash = bin2hstr(txn.txhash)
-                    js_bk1.transactions[i].pubhash = bin2hstr(txn.pubhash)
-                i += 1
+            self.reformat_block(js_bk1)
 
             return json_print_telnet(js_bk1)
 
@@ -237,17 +421,157 @@ class ApiProtocol(Protocol):
 
         return json_print_telnet(net_stats)
 
+    def search_txhash(self, txhash):  # txhash is unique due to nonce.
+        err = {'status': 'Error', 'error': 'txhash not found', 'method': 'txhash', 'parameter': txhash}
+        for tx in self.factory.chain.transaction_pool:
+            if tx.txhash == txhash:
+                logger.info('%s found in transaction pool..', txhash)
+                tx_new = copy.deepcopy(tx)
+                self.reformat_txn(tx_new)
+                return json_print_telnet(tx_new)
+
+        try:
+            txn_metadata = self.factory.chain.state.db.get(txhash)
+        except:
+            logger.info('%s does not exist in memory pool or local blockchain..', txhash)
+            return json_print_telnet(err)
+
+        json_tx = json.loads(txn_metadata[0])
+        tx = Transaction().from_txdict(json_tx)
+        tx_new = copy.deepcopy(tx)
+        self.reformat_txn(tx_new)
+        logger.info('%s found in block %s', txhash, str(txn_metadata[1]))
+        tx_new.status = 'ok'
+        return json_print_telnet(tx_new)
+
     def txhash(self, data=None):
         logger.info('<<< API tx/hash call %s', data)
-        return self.factory.chain.search_txhash(data)
+        return self.search_txhash(data)
 
     def balance(self, data=None):
         logger.info('<<< API balance call %s', data)
-        return self.factory.chain.basic_info(data)
+
+        address = data
+
+        addr = {}
+
+        if not self.factory.state.state_address_used(address):
+            addr['status'] = 'error'
+            addr['error'] = 'Address not found'
+            addr['parameter'] = address
+            return json_print_telnet(addr)
+
+        nonce, balance, _ = self.factory.state.state_get_address(address)
+        addr['state'] = {}
+        addr['state']['address'] = address
+        addr['state']['balance'] = balance / 100000000.000000000
+        addr['state']['nonce'] = nonce
+        addr['state']['transactions'] = self.factory.state.state_get_txn_count(address)
+        addr['status'] = 'ok'
+
+        return json_print_telnet(addr)
+
+    # used for port 80 api - produces JSON output reporting every transaction for an address, plus final balance..
+
+    def search_address(self, address):
+
+        addr = {'transactions': []}
+
+        txnhash_added = set()
+
+        if not self.factory.state.state_address_used(address):
+            addr['status'] = 'error'
+            addr['error'] = 'Address not found'
+            addr['parameter'] = address
+            return json_print_telnet(addr)
+
+        nonce, balance, pubhash_list = self.factory.state.state_get_address(address)
+        addr['state'] = {}
+        addr['state']['address'] = address
+        addr['state']['balance'] = balance / 100000000.000000000
+        addr['state']['nonce'] = nonce
+
+        for s in self.factory.state.stake_list_get():
+            if address == s[0]:
+                addr['stake'] = {}
+                addr['stake']['selector'] = s[2]
+                # pubhashes used could be put here..
+
+        tmp_transactions = []
+        for tx in self.factory.chain.transaction_pool:
+            if tx.subtype not in (TX_SUBTYPE_TX, TX_SUBTYPE_COINBASE):
+                continue
+            if tx.txto == address or tx.txfrom == address:
+                logger.info('%s found in transaction pool', address)
+
+                tmp_txn = {'subtype': tx.subtype,
+                           'txhash': bin2hstr(tx.txhash),
+                           'block': 'unconfirmed',
+                           'amount': tx.amount / 100000000.000000000,
+                           'nonce': tx.nonce,
+                           'ots_key': tx.ots_key,
+                           'txto': tx.txto,
+                           'txfrom': tx.txfrom,
+                           'timestamp': 'unconfirmed'}
+
+                if tx.subtype == TX_SUBTYPE_TX:
+                    tmp_txn['fee'] = tx.fee / 100000000.000000000
+
+                tmp_txn.subtype = Transaction.tx_id_to_name(tx.subtype)
+
+                tmp_transactions.append(tmp_txn)
+                txnhash_added.add(tx.txhash)
+
+        addr['transactions'] = tmp_transactions
+
+        my_txn = []
+        try:
+            my_txn = self.factory.state.db.get('txn_' + address)
+        except:
+            pass
+
+        for txn_hash in my_txn:
+            txn_metadata = self.factory.state.db.get(txn_hash)
+            dict_txn_metadata = json.loads(txn_metadata[0])
+            if dict_txn_metadata['subtype'] == TX_SUBTYPE_TX:
+                tx = SimpleTransaction().json_to_transaction(txn_metadata[0])
+            elif dict_txn_metadata['subtype'] == TX_SUBTYPE_COINBASE:
+                tx = CoinBase().json_to_transaction(txn_metadata[0])
+
+            if (tx.txto == address or tx.txfrom == address) and tx.txhash not in txnhash_added:
+                logger.info('%s found in block %s', address, str(txn_metadata[1]))
+
+                tmp_txn = {'subtype': tx.subtype,
+                           'txhash': bin2hstr(tx.txhash),
+                           'block': txn_metadata[1],
+                           'timestamp': txn_metadata[2],
+                           'amount': tx.amount / 100000000.000000000,
+                           'nonce': tx.nonce,
+                           'ots_key': tx.ots_key,
+                           'txto': tx.txto,
+                           'txfrom': tx.txfrom}
+
+                if tx.subtype == TX_SUBTYPE_TX:
+                    tmp_txn['fee'] = tx.fee / 100000000.000000000
+
+                tmp_txn['subtype'] = Transaction.tx_id_to_name(tmp_txn['subtype'])
+
+                addr['transactions'].append(tmp_txn)
+                txnhash_added.add(tx.txhash)
+
+        if len(addr['transactions']) > 0:
+            addr['state']['transactions'] = len(addr['transactions'])
+
+        if addr == {'transactions': {}}:
+            addr = {'status': 'error', 'error': 'address not found', 'method': 'address', 'parameter': address}
+        else:
+            addr['status'] = 'ok'
+
+        return json_print_telnet(addr)
 
     def address(self, data=None):
         logger.info('<<< API address call %s', data)
-        return self.factory.chain.search_address(data)
+        return self.search_address(data)
 
     def dataReceived(self, data=None):
         self.parse_cmd(data)
